Update files and things

+++ docker-entrypoint.sh
@@ -4,6 +4,11 @@
 #  Docker Entrypoint for Forem
 #
 
+if [[ "$1" == "bash" ]]; then
+    bash
+    exit 0
+fi
+
 set -e
 
 # Exit
@@ -26,15 +31,27 @@ BG_CYAN="\033[46m"
 BG_RED="\033[41m"
 BG_YELLOW="\033[43m"
 
+# Initialize forem
+
+[[ ! -f "/forem/.copied" ]] && rm -rf /forem/*
+[[ ! -f "/forem/.copied" ]] && rm -rf /forem/.* || true
+if [[ ! -f "/forem/.copied" ]]; then
+    for f in $(ls -A /forem-tmp); do
+        cp -r "/forem-tmp/${f}" /forem
+    done
+fi
+[[ ! -f "/forem/.copied" ]] && rm -rf /forem-tmp
+[[ ! -f "/forem/.copied" ]] && touch /forem/.copied
+
 # Initialize nvm and rbenv
 
 printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Initializing programs...${RESET}\n"
 
-source $HOME/.bashrc
+source "${HOME}/.bashrc"
 export PATH="$HOME/.rbenv/bin:$PATH"
 eval "$(rbenv init -)"
 export PATH="$HOME/.rbenv/plugins/ruby-build/bin:$PATH"
-. "$S_NVM_DIR/nvm.sh"
+. "${S_NVM_DIR}/nvm.sh"
 
 # Set variables
 
@@ -82,23 +99,18 @@ sed -i "s/secret_key_base: <%= ENV\[\"SECRET_KEY_BASE\"\] %>/secret_key_base: ${
 
 # Prepare services
 
-printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Changing permissions on PostgreSQL, Redis, and ElasticSearch data directories...${RESET}\n"
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Initializing PostgreSQL...${RESET}\n"
 
-chown -R postgres:postgres /var/lib/postgresql/13/main
-chmod -R a+rwx /var/lib/postgresql/13/main
-chown -R root:root /var/lib/redis
-chown -R root:root /var/lib/elasticsearch
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres sudo rm -rf /var/lib/postgresql/13/main/*
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres sudo rm -rf /var/lib/postgresql/13/main/.* || true
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres /usr/lib/postgresql/13/bin/initdb --encoding=UTF8 -D /var/lib/postgresql/13/main
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && touch "/var/lib/postgresql/13/main/.configured"
 
-printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Initializing PostgreSQL...${RESET}\n"
+[[ -d "/elastic-temp" ]] && printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Initializing ElasticSearch...${RESET}\n"
 
-chmod -R a+rwx /var/lib/postgresql/13/main
-sudo -u postgres /usr/lib/postgresql/13/bin/initdb -D /var/lib/postgresql/13/main
-
-printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Initializing ElasticSearch...${RESET}\n"
-
-cp -r /elastic-temp/* /var/lib/elasticsearch
-chmod -R a+rwx /var/lib/elasticsearch
-rm -r /elastic-temp
+[[ -d "/elastic-temp" ]] && cp -r /elastic-temp/* /var/lib/elasticsearch
+[[ -d "/elastic-temp" ]] && chmod -R a+rwx /var/lib/elasticsearch
+[[ -d "/elastic-temp" ]] && rm -r /elastic-temp
 
 # Start services
 
@@ -112,34 +124,36 @@ service elasticsearch start
 
 printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Creating PostgreSQL user...${RESET}\n"
 
-sudo -u postgres createuser -s ${S_POSTGRES_USER}
-sudo -u postgres psql -c "ALTER USER ${S_POSTGRES_USER} WITH PASSWORD '${S_POSTGRES_PASSWORD}';"
-sudo -u postgres psql -c "CREATE DATABASE ${S_POSTGRES_USER};"
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres createuser -s "${S_POSTGRES_USER}"
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres psql -c "ALTER USER ${S_POSTGRES_USER} WITH PASSWORD '${S_POSTGRES_PASSWORD}';"
+[[ ! -f "/var/lib/postgresql/13/main/.configured" ]] && sudo -u postgres psql -c "CREATE DATABASE ${S_POSTGRES_USER};"
 
 # Get more environment variables
 
 printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Checking other environment variables...${RESET}\n"
 
-source /forem/.env
+source "/forem/.env"
 
 # Configure forem
 
 printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Checking for development environment...${RESET}\n"
 
-FOREM_LOWERCASE_ENV=$(echo $FOREM_ENV | tr '[:upper:]' '[:lower:]')
+FOREM_LOWERCASE_ENV="$(echo "${FOREM_ENV}" | tr '[:upper:]' '[:lower:]')"
 
-if [[ "$FOREM_LOWERCASE_ENV" == "production" ]]; then
+if [[ "$FOREM_LOWERCASE_ENV" == "production" ]] && [[ ! -f "/forem/.ready" ]]; then
     printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Environment is production.${RESET}\n"
 
     # Configure Forem
 
     printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Configuring forem...${RESET}\n"
 
+    bin/rails db:create
     bin/setup
 
     printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Compiling webpack assets...${RESET}\n"
 
     node /webpack-config.js
+    bin/webpack
 
     # Clear database
 
@@ -161,6 +175,8 @@ if [[ "$FOREM_LOWERCASE_ENV" == "production" ]]; then
     printf "${BLACK}${BG_CYAN} INFO ${RESET} ${BLUE}Reconfiguring forem...${RESET}\n"
 
     bin/setup
+
+    [[ ! -f "/forem/.ready" ]] && touch "/forem/.ready"
 else
     printf "${BLACK}${BG_YELLOW} WARN ${RESET} ${YELLOW}Environment is development.${RESET}\n"
 fi
